# medical-project-backend/backend/app/static/openapi.yaml
openapi: 3.1.0
info:
  title: Patient Management API
  version: "1.2.0"
  description: >
    API for managing patients, including creating new patients, updating existing patient records (with version history), medical history, medical visits, and retrieving detailed patient information.
servers:
  - url: https://8dj6t53su8wv.share.zrok.io
    description: Local development server
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate a user during the login process
      description: >
        Verify that the credentials given match the hash in the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
            examples:
              Example1:
                summary: Sample Authentication
                value:
                  email: "john.doe@example.com"
                  password: "mysecretpassword"
      responses:
        '200':
          description: Authentication successful or MFA required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                SuccessResponse:
                  summary: Successful login
                  value:
                    status: "Success"
                    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    user:
                      user_id: "99febb18-8a4c-40b6-942f-df3c60d522dd"
                      user_type: "PATIENT" 
                      name: "John Doe"
                      medical_insurance_id: "INS123456"
                      requires_mfa: false
                MfaRequiredResponse:
                  summary: MFA verification required
                  value:
                    status: "MFA Required"
                    temp_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    user:
                      user_id: "99febb18-8a4c-40b6-942f-df3c60d522dd"
                      user_type: "PATIENT" 
                      name: "John Doe"
                      medical_insurance_id: "INS123456"
                      requires_mfa: true
        '401':
          description: Unsuccessful login attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                WrongCredentialsResponse:
                  summary: Unsuccessful login
                  value:
                    status: "Wrong credentials"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                ServerErrorResponse:
                  summary: Internal Server Error
                  value:
                    status: "Internal server error"
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout a user by blacklisting their token
      description: >
        Revokes the provided JWT token to prevent its further use.
        Requires authentication through a valid token.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "Successfully logged out"
              examples:
                SuccessResponse:
                  summary: Successful logout
                  value:
                    status: "Successfully logged out"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                InvalidTokenResponse:
                  summary: Invalid token
                  value:
                    status: "error"
                    error: "Invalid token"
                    message: "Signature verification failed"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ServerErrorResponse:
                  summary: Internal Server Error
                  value:
                    error: "Error during logout"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
...
    ErrorResponse:
      type: object
      description: Schema for error responses.
      properties:
        error:
          type: string
          example: "Invalid foreign key reference."
      required:
        - error
      additionalProperties: false
    Login:
      type: object
      description: Schema for authenticating a user.
      properties:
        email:
          type: string
          format: email
          example: "john.doe@example.com"
          description: User's email address.
        password:
          type: string
          description: Plain text password. It will be hashed after submission.
          example: "mysecretpassword"
      required:
        - email
        - password
      additionalProperties: false
...
    LoginResponse:
      type: object
      description: Schema for login responses.
      properties:
        status:
          type: string
          example: "Success"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        temp_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          type: object
          properties:
            user_id:
              type: string
              format: uuid
              example: "99febb18-8a4c-40b6-942f-df3c60d522dd"
            user_type:
              type: string
              example: "PATIENT"
            name:
              type: string
              example: "John Doe"
            medical_insurance_id:
              type: string
              example: "INS123456"
            requires_mfa:
              type: boolean
              example: false
      required:
        - status
...